Cast int to size_t

+++ fgk/utlities.c
@@ -394,7 +394,7 @@ char* get_filename(char* string){
             last_slash = i;
         }
     }
-    char* output = calloc(length, sizeof(char));
+    char* output = calloc((size_t) length, sizeof(char));
 
     if(last_slash == -1){
         strcpy(output, string);
